+++ Aquaria/Continuity.cpp
@@ -2378,7 +2378,7 @@ void Continuity::saveFile(int slot, Vector position, unsigned char *scrShotData,
 	XMLElement *worldMap = doc.NewElement("WorldMap");
 	{
 		std::ostringstream os;
-		for (int i = 0; i < dsq->continuity.worldMap.getNumWorldMapTiles(); i++)
+		for (size_t i = 0; i < dsq->continuity.worldMap.getNumWorldMapTiles(); i++)
 		{
 			WorldMapTile *tile = dsq->continuity.worldMap.getWorldMapTile(i);
 			if (tile->revealed)
@@ -2391,7 +2391,7 @@ void Continuity::saveFile(int slot, Vector position, unsigned char *scrShotData,
 		if (dsq->game->worldMapRender)
 		{
 			std::ostringstream os;
-			for (int i = 0; i < dsq->continuity.worldMap.getNumWorldMapTiles(); i++)
+			for (size_t i = 0; i < dsq->continuity.worldMap.getNumWorldMapTiles(); i++)
 			{
 				WorldMapTile *tile = dsq->continuity.worldMap.getWorldMapTile(i);
 				os << tile->index << " ";
